+++ src/com/jeu/main/HUD.java
@@ -5,15 +5,15 @@ import java.awt.Graphics;
 
 public class HUD {
 	
-	public static int PV = 100;
-	private int greenValue = 255;
+	public static float PV = 100;
+	private float greenValue = 255;
 	
 	private int score = 0;
 	private int niveau = 1;
 
 	public void tick() {		
-		PV = Game.clamp(PV, 0, 100);
-		greenValue = Game.clamp(greenValue, 0, 255);
+		PV = (int) Game.clamp(PV, 0, 100);
+		greenValue = (int) Game.clamp(greenValue, 0, 255);
 		greenValue = PV * 2;
 		score++;
 	}
@@ -21,8 +21,8 @@ public class HUD {
 	public void render(Graphics graphique) {
 		graphique.setColor(Color.gray);
 		graphique.fillRect(15, 15, 200, 32);
-		graphique.setColor(new Color(75, greenValue, 0));
-		graphique.fillRect(15, 15, PV * 2, 32);
+		graphique.setColor(new Color(75, (int) greenValue, 0));
+		graphique.fillRect(15, 15, (int) PV * 2, 32);
 		graphique.setColor(Color.white);
 		graphique.drawRect(15, 15, 200, 32);
